+++ galeria-app/app/src/main/java/com/example/vhra/galeria/presenter/Presenter.java
@@ -2,7 +2,12 @@ package com.example.vhra.galeria.presenter;
 
 import com.example.vhra.galeria.view.adapter.MediasAdapter;
 import com.example.vhra.galeria.model.Media;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Locale;
 
 public class Presenter {
 
@@ -10,12 +15,14 @@ public class Presenter {
 
     private List<Media> medias;
 
+    private static final List<String> FILE_EXTENSIONS =  Arrays.asList("jpg", "jpeg", "png");
+
     public Presenter(IPresenter view) {
         this.view = view;
     }
 
     public void onCreate() {
-        this.medias = Media.findAll();
+        this.medias = findAllMedia();
 
         if (view != null)
             view.setAdapter(new MediasAdapter( medias ));
@@ -25,4 +32,25 @@ public class Presenter {
         view = null;
     }
 
+
+    private List<Media> findAllMedia() {
+        List<Media> medias = new ArrayList<>();
+
+        File directory = new File(android.os.Environment.getExternalStorageDirectory() + File.separator  + "Download");
+        if (directory.isDirectory()) {
+            for (File file : directory.listFiles()) {
+                String filePath = file.getAbsolutePath();
+                if (isSupportedFile(filePath)) {
+                    medias.add(new Media(filePath));
+                }
+            }
+        }
+
+        return medias;
+    }
+
+    private static boolean isSupportedFile(String filePath) {
+        String ext = filePath.substring((filePath.lastIndexOf(".") + 1), filePath.length());
+        return FILE_EXTENSIONS.contains(ext.toLowerCase(Locale.getDefault()));
+    }
 }
